pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-cloud-storage>=3.4.0",
    "gunicorn>=23.0.0",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "werkzeug>=3.1.3",
]
